fix: change .chain() to .sources() to use new Rust (#391)
* fix: change .chain() to .sources() and change CI

* fix: map_first_last feature is now stable

* chore: supress clippy errors

* fix: fixed typo

* chore: cargo fmt and clippy

* chore: clippy

+++ synth/src/sampler.rs
@@ -34,7 +34,7 @@ impl SamplerOutput {
 }
 
 fn sampler_progress_bar(target: u64) -> ProgressBar {
-    let bar = ProgressBar::new(target as u64);
+    let bar = ProgressBar::new(target);
     let style = ProgressStyle::default_bar()
         .template("[{elapsed_precise}] {wide_bar} {pos}/{len} generated ({eta} remaining)");
     bar.set_style(style);
